requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.1
transformers==4.38.1
